Cleaned up store creation a bit
